Added Snake2
Added Snake2.
Snake two is a more object oriented snake build which is only 261 lines of original code. I could probably get it down a bit with some more optimization but im happy with where it is.

Interesting thing i learned in this project was that the snake is very similar to a queue, as long as you drop the notion of a static head and tail. Using a queue, i only need to add or delete a maximum of two snake parts(+head -tail) at any given interval.